requirements.txt
requests
python-telegram-bot[job-queue]
Flask
gunicorn
psycopg2-binary